package: introduce virtual package concept
Introduce a special supported _virtual package format. This is
especially designed to handle future packages that don't exist in Rift
project. Instances of concrete PackageRPM package class are returned by
ProjectPackages only when the package actually exist in the project.

This is especially useful to manage packages removed in patches.