Support static analysis of Containerfile
Rift now supports static analysis of Containerfile by calling hadolint
linter. This includes rift check command, rift gerrit command and
automatic check on validation.

In the same way Spec class handles the RPM package source file, the
ContainerFile class is introduced to handle operations on OCI container
source definition file.

Hadolint is a weak dependency, it is used when present on system but
Rift does not fail if not. It just emits an error log if executable
cannot be found.

Tests are added to run tests with hadolint integration, tests are
gently skipped if hadolint executable cannot be found.